+++ src/main/java/io/myzticbean/finditemaddon/config/ConfigProvider.java
@@ -21,6 +21,7 @@ package io.myzticbean.finditemaddon.config;
 import io.myzticbean.finditemaddon.utils.log.Logger;
 import me.kodysimpson.simpapi.colors.ColorTranslator;
 import org.bukkit.Bukkit;
+import org.bukkit.Material;
 import org.bukkit.World;
 
 import java.util.ArrayList;
@@ -39,6 +40,7 @@ public class ConfigProvider {
     private static final String SHOP_GUI = "shop-gui.";
     private static final String SHOP_GUI_NAVIGATION = "shop-navigation.";
     private static final String SHOP_GUI_CMD = "custom-model-data.";
+    private static final String BENTOBOX = "bentobox.";
     public final String PLUGIN_PREFIX = ColorTranslator.translateColorCodes(ConfigSetup.get().getString("plugin-prefix"));
     public final List<String> FIND_ITEM_COMMAND_ALIAS = (List<String>) ConfigSetup.get().getList(FIND_ITEM_COMMAND + "command-alias");
     public final String FIND_ITEM_TO_BUY_AUTOCOMPLETE = ConfigSetup.get().getString(FIND_ITEM_COMMAND + "to-buy-autocomplete");
@@ -100,26 +102,28 @@ public class ConfigProvider {
     public final boolean SHOP_GUI_USE_SHORTER_CURRENCY_FORMAT = ConfigSetup.get().getBoolean(SHOP_GUI + "use-shorter-currency-format");
     public final int SHOP_PLAYER_VISIT_COOLDOWN_IN_MINUTES = ConfigSetup.get().getInt("shop-player-visit-cooldown-in-minutes");
     public final boolean IGNORE_EMPTY_CHESTS = ConfigSetup.get().getBoolean("ignore-empty-chests");
-    public final List<String> BLACKLISTED_WORLDS = (List<String>) ConfigSetup.get().getList("blacklisted-worlds");
+    private final List<String> BLACKLISTED_WORLDS = (List<String>) ConfigSetup.get().getList("blacklisted-worlds");
+    private final List<String> BLACKLISTED_MATERIALS = (List<String>) ConfigSetup.get().getList("blacklisted-materials");
+    public final boolean SUPPRESS_UPDATE_NOTIFICATIONS = ConfigSetup.get().getBoolean("suppress-update-notifications");
     public final boolean DEBUG_MODE = ConfigSetup.get().getBoolean("debug-mode");
     public final int CONFIG_VERSION = ConfigSetup.get().getInt("config-version");
+    public final boolean BENTOBOX_IGNORE_LOCKED_ISLAND_SHOPS = ConfigSetup.get().getBoolean(BENTOBOX + "ignore-locked-island-shops");
 
     private final List<World> blacklistedWorldsList = new ArrayList<>();
+    private final List<Material> blacklistedMaterialsList = new ArrayList<>();
 
     public ConfigProvider() {
-        if(BLACKLISTED_WORLDS != null) {
-            BLACKLISTED_WORLDS.forEach(world -> {
-                World worldObj = Bukkit.getWorld(world);
-                if(worldObj != null) {
-                    blacklistedWorldsList.add(worldObj);
-                }
-            });
-        }
+        loadBlacklistedWorlds();
+        loadBlacklistedMaterials();
         Logger.logInfo("Config loaded!");
     }
 
     public List<World> getBlacklistedWorlds() {
         return blacklistedWorldsList;
+    }
+
+    public List<Material> getBlacklistedMaterials() {
+        return blacklistedMaterialsList;
     }
 
     public boolean shopGUIItemLoreHasKey(String key) {
@@ -128,6 +132,28 @@ public class ConfigProvider {
         }
         else {
             return false;
+        }
+    }
+
+    private void loadBlacklistedWorlds() {
+        if(BLACKLISTED_WORLDS != null) {
+            BLACKLISTED_WORLDS.forEach(world -> {
+                World worldObj = Bukkit.getWorld(world);
+                if(worldObj != null) {
+                    blacklistedWorldsList.add(worldObj);
+                }
+            });
+        }
+    }
+
+    private void loadBlacklistedMaterials() {
+        if(BLACKLISTED_MATERIALS != null) {
+            BLACKLISTED_MATERIALS.forEach(material -> {
+                var materialObj = Material.getMaterial(material);
+                if(materialObj != null) {
+                    blacklistedMaterialsList.add(materialObj);
+                }
+            });
         }
     }
 }